# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "开源、自托管轻量级发布平台 Ech0 的 API 文档",
        "title": "Ech0 API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6277",
    "basePath": "/api",
    "paths": {
        "/allusers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取系统中所有用户的详细信息，接口需要认证",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取所有用户",
                "responses": {
                    "200": {
                        "description": "获取失败，code=0，msg错误描述",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用户通过用户名和密码登录，返回 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录接口",
                "parameters": [
                    {
                        "description": "登录请求体",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录失败，返回错误信息",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "通过提交用户名、密码等信息完成注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求体",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误或注册失败，code=0，msg错误描述",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前认证用户的详细信息，密码字段不会返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取当前用户信息",
                "responses": {
                    "200": {
                        "description": "获取失败，code=0，msg错误描述",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "接口会根据请求体更新用户相关字段，需携带有效的用户身份（如 JWT）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新当前用户的信息",
                "parameters": [
                    {
                        "description": "用户更新信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserInfoDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误或更新失败，code=0，msg错误描述",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/admin/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "通过用户ID更新其管理员权限，接口调用者需拥有相应权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新用户权限（管理员权限）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或更新失败，code=0，msg错误描述",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据用户ID删除用户，调用者需具备相应权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或删除失败，code=0，msg错误描述",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 状态码，非0时表示自定义HTTP业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 响应数据，具体内容因接口而异"
                },
                "msg": {
                    "description": "Msg 返回信息，通常是状态描述",
                    "type": "string"
                }
            }
        },
        "model.LoginDto": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.RegisterDto": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.UserInfoDto": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像地址\nexample: https://example.com/avatar.png",
                    "type": "string"
                },
                "is_admin": {
                    "description": "是否为管理员\nexample: false",
                    "type": "boolean"
                },
                "password": {
                    "description": "密码\nexample: 123456",
                    "type": "string"
                },
                "username": {
                    "description": "用户名\nexample: linsnow",
                    "type": "string"
                }
            }
        }
    }
}